requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
argon2-cffi==23.1.0
Pillow==10.2.0